requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
cohere==5.3.3
qdrant-client==1.9.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
tiktoken==0.7.0
langchain-text-splitters==0.2.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
openai==1.3.4
httpx==0.25.2
